Book Reader Lab (zipped)
The purpose of this lab was to write a program that reads public domain books from .txt files (the three provided materials were titled alice, siddhartha, and odyssey). It then returns and displays the number of unique words in the chosen book and, if prompted, displays a list with the numbers of repetitions of each unique word.

EXAMPLE OUTPUT:

Please enter book title/filepath (or enter q to quit): siddhartha

There are 3582 unique words in Siddhartha.
Show # of repetitions of each unique word in Siddhartha? Y/N: y

Frequency of Each Unique Word in SIDDHARTHA:
====================================================
1x  'nicely'
6x  'half'
61x  'spoke'
(and etc; it's a very long list)